requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.7.1
lxml==5.3.0
numpy==2.1.3
openai==1.54.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
playwright==1.48.0
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1